NRC7292 Standalone SDK v1.3.2 release

+++ package/standalone/lib/hostap/src/drivers/driver_nrc_scan.h
@@ -52,7 +52,7 @@ void scan_start(struct nrc_scan_info *scan);
 void scan_stop(struct nrc_scan_info *scan);
 void scan_cancel(struct nrc_scan_info *scan);
 uint16_t scan_resume_from(struct nrc_scan_info *scan);
-int scan_add(struct nrc_scan_info *scan, uint16_t freq, uint16_t rssi,
+int scan_add(struct nrc_scan_info *scan, uint16_t freq, int8_t rssi,
 		uint8_t* frame, uint16_t len);
 void scan_config(struct nrc_scan_info *scan, struct wpa_driver_scan_params *p,
 		uint16_t last_freq);
